% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baecontemplate-package.R
\docType{package}
\name{baecontemplate-package}
\alias{baecontemplate}
\alias{baecontemplate-package}
\title{baecontemplate: pkgdown Templates for baeconverse Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This is a pkgdown template for baeconverse packages.
}
\author{
\strong{Maintainer}: Young Geun Kim \email{ygeunkimstat@gmail.com} (\href{https://orcid.org/0000-0001-8651-1167}{ORCID}) [copyright holder]

}
\keyword{internal}
